update dpo_trainer dpo_train and readme
